renaming utilities -> utils
