+++ docs/api/index.rst
@@ -2,7 +2,7 @@ Auxjad API
 ==========
 
 Auxjad is made out of nine subpackages: |core|_, |get|_, |indicators|_,
-|makers|_, |mutate|_, |score|_, |select|_, |spanners|_, and |utilities|_.
+|makers|_, |mutate|_, |score|_, |select|_, |spanners|_, and |utils|_.
 
 |core|_ contains most of Auxjad's classes and is focused on algorithmic
 transformations and manipulations of |abjad.Container| objects. |indicators|_
@@ -10,7 +10,7 @@ and |spanners|_ contain derived classes and extension methods for Abjad's
 indicators and spanners. |score|_ contains score component classes, such as
 harmonic leaves. |makers|_ contain an expanded leaf making class. |get|_,
 |mutate|_, and |select|_ contain inspection, selection, and mutation functions,
-respectively. |utilities|_ contain general utility functions..
+respectively. |utils|_ contain general utility functions..
 
 ..  note::
 
@@ -42,7 +42,7 @@ in the navigation pane in the left.
 +----------------+-------------------------------------------------------------+
 | |spanners|_    | spanner classes                                             |
 +----------------+-------------------------------------------------------------+
-| |utilities|_   | utility functions                                           |
+| |utils|_       | utility functions                                           |
 +----------------+-------------------------------------------------------------+
 
 .. toctree::
@@ -58,7 +58,7 @@ in the navigation pane in the left.
     api-score
     api-select
     api-spanners
-    api-utilities
+    api-utils
 
 .. include:: abjad-targets.rst
 .. include:: auxjad-targets.rst
