+++ src/auxjad/utils/staff_splitter.py
